refactor: domain에 vo 패키지 추가
