+++ backend/src/main/java/team/teamby/teambyteam/teamplace/domain/TeamPlace.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import team.teamby.teambyteam.member.domain.MemberTeamPlace;
 import team.teamby.teambyteam.member.domain.vo.Email;
+import team.teamby.teambyteam.teamplace.domain.vo.Name;
 
 import java.util.ArrayList;
 import java.util.List;
