+++ backend/src/test/java/team/teamby/teambyteam/schedule/domain/TitleTest.java
@@ -2,6 +2,7 @@ package team.teamby.teambyteam.schedule.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import team.teamby.teambyteam.schedule.domain.vo.Title;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
